Extract and centralize test utilities for reusable testing across modules
