+++ publishing/digital-ocean-spaces/build.gradle.kts
@@ -1,6 +1,5 @@
 val publishingDigitalOceanSpacesVersion: String by rootProject.extra
 
-
 plugins {
     `kotlin-dsl`
     id("org.jetbrains.dokka") version "1.9.20"
@@ -30,6 +29,8 @@ repositories {
 dependencies {
     implementation("software.amazon.awssdk:s3:2.25.27")
 
+    testImplementation(gradleTestKit())
+    testImplementation(project(":test-core"))
     testImplementation("org.jetbrains.kotlin:kotlin-test-junit5")
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
     testImplementation("io.mockk:mockk:1.13.8")
