+++ publishing/digital-ocean-spaces/src/test/kotlin/io/violabs/plugins/open/publishing/digitalocean/FunctionalPluginStandaloneTest.kt
@@ -1,0 +1,6 @@
+package io.violabs.plugins.open.publishing.digitalocean
+
+import io.violabs.test.core.UnitTest
+
+class FunctionalPluginStandaloneTest : UnitTest() {
+}
