+++ publishing/digital-ocean-spaces/src/test/resources/functional/settings.gradle.test.txt
@@ -1,0 +1,1 @@
+rootProject.name = "test"
