+++ publishing/maven-generated-artifacts/src/test/kotlin/io/violabs/plugins/open/publishing/mavengenerated/PluginServiceFunctionalTest.kt
@@ -1,5 +1,6 @@
 package io.violabs.plugins.open.publishing.mavengenerated
 
+import io.violabs.test.core.UnitTest
 import org.gradle.testkit.runner.GradleRunner
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
