+++ test-core/src/main/kotlin/io/violabs/test/core/UnitTest.kt
@@ -1,11 +1,6 @@
-package io.violabs.plugins.open.publishing.mavengenerated
-
-import org.gradle.api.logging.Logger
-import org.gradle.api.logging.Logging
+package io.violabs.test.core
 
 abstract class UnitTest {
-    protected val logger: Logger = Logging.getLogger(UnitTest::class.java)
-
     protected fun <A> testEquals(expected: A, actual: A): Boolean {
         if (expected == actual) return true
